embeddings/requirements.txt
fastapi
uvicorn[standard]
pillow
tensorflow
numpy
python-multipart  # Needed to handle file uploads



